requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
pydantic==1.8.2
pandas==1.3.3
pymongo==3.12.0
scikit-learn==0.24.2
numpy==1.21.2
scipy==1.7.1
tensorflow==2.5.0
Pillow==8.3.1
protobuf==3.20.3
pymongo[srv]==3.12.0
